Rename component folders to have container
