+++ src/pages/RestaurantsPage/CategoriesContainer/CategoriesContainer.jsx
