+++ src/pages/RestaurantsPage/CategoriesContainer/CategoriesContainer.test.jsx
