+++ src/pages/RestaurantsPage/CategoriesContainer/index.js
